Move config.ts to src/server/ so that it gets formatted and linted
